fix(frontend): show curl|sh install command on settings/daemons for offline reconnect
Daemons settings page had no actionable info when a daemon went offline. Lift
the curl|sh install panel out of CreateAgentDialog into a shared
DaemonInstallCommand component and render it on /settings/daemons when the
list is empty or contains offline daemons. Same install-ticket flow as Create
Agent — same machine reuses creds, new machines auto-authorize.
